Address PR #1611 review comments
- Restore package-lock.json libc metadata accidentally stripped by the branch
- Document why BallotActionType currently mirrors BallotSubmitType
- Fix encodeBallotRow test to also assert race2Vote when all races are missing
- Reuse testInputs.Election1 as the base for allowedSubmitTypes.test's makeElection
- Move electionSettingsValidation unit tests from the backend to the shared
  package, next to the function they test, adding vitest as its test runner
- Switch frontend unit tests from jest to vitest, since frontend never
  declared its own jest/ts-jest dependency (it only worked via hoisting
  from backend's devDependencies)
- Fix lint-staged failures uncovered while touching these files: unused
  imports/empty interface in Ballot.ts, and a require()-style dotenv import
  in allowedSubmitTypes.test.ts
