+++ packages/frontend/package.json
@@ -37,7 +37,7 @@
     "build": "tsc && rsbuild build",
     "start": "rsbuild preview",
     "lint": "eslint . --fix",
-    "test": "jest --config jest.config.cjs"
+    "test": "vitest run"
   },
   "browserslist": {
     "production": [
@@ -69,6 +69,7 @@
     "globals": "^16.0.0",
     "rimraf": "^5.0.5",
     "typescript": "^5.9.0",
-    "typescript-eslint": "^8.30.1"
+    "typescript-eslint": "^8.30.1",
+    "vitest": "^5.0.0"
   }
 }
