--- packages/frontend/tsconfig.test.json
@@ -1,11 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "lib": ["ES2022"],
-    "types": ["jest", "node"],
-    "isolatedModules": true
-  },
-  "exclude": []
-}
