+++ packages/frontend/vitest.config.ts
@@ -1,0 +1,13 @@
+import { defineConfig } from 'vitest/config';
+import path from 'node:path';
+
+export default defineConfig({
+    resolve: {
+        alias: {
+            '~': path.resolve(import.meta.dirname, './src'),
+        },
+    },
+    test: {
+        include: ['src/**/*.test.ts', 'src/**/*.test.tsx'],
+    },
+});
