+++ packages/shared/tsconfig.json
@@ -18,5 +18,6 @@
     "declarationMap": true,
     "declarationDir": "./dist/types"
   },
-  "include": ["./src/**/*"]
+  "include": ["./src/**/*"],
+  "exclude": ["./src/**/*.test.ts"]
 }
